requirements.txt
apache-airflow>=2.9.0
apache-airflow-providers-http

apache-airflow-providers-common-sql>=1.7.2

requests
pandas
SQLAlchemy
flask<2.3,>=2.2
flask-appbuilder==4.5.0
python-slugify>=7.0.0
click<9.0,>=8.1.2
attrs>=22.2.0
opentelemetry-proto==1.25.0

apache-airflow[amazon] 
apache-airflow-providers-amazon 
boto3
requests 

pyspark